src/PayChain_backend_python/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
sqlalchemy==1.4.52
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
databases[sqlite]==0.8.0
aiosqlite==0.19.0
python-dotenv==1.0.0
email-validator==2.2.0 